Adjust Marshal/Unmarshal Interface (#10)
Previously the Unmarshal function wrapper for the codec was passing interface directly not as a reference which was a bit non-standard and would cause a bit overhead. Also the marshal wrapper had the issue of returning a string, which could cause the need of additional conversions in use-cases external to this cli tool.